move build.rs code to binaries.rs
