--- fetcher-core/src/build.rs
@@ -1,363 +1,0 @@
-use reqwest::{Certificate, Url};
-use sha2::{Digest, Sha512};
-use std::{
-    env,
-    env::var,
-    fs,
-    fs::File,
-    io::{BufRead, BufReader, Read, Write},
-    os::unix::fs::OpenOptionsExt,
-    path::Path,
-    path::PathBuf,
-    process::{Command, Stdio},
-    time::Duration,
-};
-use tokio::task::JoinSet;
-
-fn main() {
-    // println!("FETCHER build.rs running");
-    // generate_config_file();
-    // binaries_main();
-}
-
-fn get_out_dir() -> PathBuf {
-    PathBuf::from(env::var("OUT_DIR").expect("OUT_DIR to be defined"))
-}
-
-fn get_manifest_dir() -> PathBuf {
-    PathBuf::from(var("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR to be set"))
-}
-
-fn get_hashsums_dir() -> PathBuf {
-    get_manifest_dir().join("hashsums")
-}
-
-fn generate_config_file() {
-    let out_dir = get_out_dir();
-
-    // Create the generated config file
-    let config_file_path = out_dir.join("config.rs");
-    let mut file = File::create(&config_file_path).expect("config.rs to be created");
-    file.write_fmt(core::format_args!(
-        "pub const FETCHER_OUT_DIR: &str = {:?};",
-        out_dir.display()
-    ))
-    .expect("config file to be written")
-}
-
-#[tokio::main]
-pub async fn binaries_main() {
-    // find or fetch zingo-blessed binaries.
-    let mut seek_binaries: JoinSet<()> = JoinSet::new();
-
-    let bin_names = vec![
-        "lightwalletd",
-        "zainod",
-        "zcashd",
-        "zcash-cli",
-        "zebrad",
-        "zingo-cli",
-    ];
-
-    for n in bin_names {
-        // Client uses an Arc internally.
-        seek_binaries.spawn(validate_binary(n));
-    }
-
-    seek_binaries.join_all().await;
-    println!("program exiting, declare victory!");
-}
-
-async fn validate_binary(binary_name: &str) {
-    let resources_dir: PathBuf = get_out_dir();
-    let bin_dir = Path::new(&resources_dir).join("test_binaries");
-    let bin_path = bin_dir.join(binary_name);
-    let shasum_path = get_hashsums_dir().join(format!("{}_shasum", binary_name));
-
-    fs::create_dir_all(bin_dir).expect("bin directory to be created");
-
-    loop {
-        if !bin_path.is_file() {
-            println!("{:?} = file not found!", &bin_path);
-            // we have to go get it!
-            fetch_binary(&bin_path, binary_name).await
-        }
-        if bin_path.is_file() {
-            //file is found, perform checks
-            match confirm_binary(&bin_path, &shasum_path, binary_name).await {
-                Ok(()) => {
-                    println!("{} binary confirmed.", &binary_name);
-                    break;
-                }
-                _ => println!(
-                    "binary {} confirmation failure, deleted found binary. plz fetch again",
-                    &binary_name
-                ),
-            }
-        }
-    }
-}
-
-async fn confirm_binary(
-    bin_path: &PathBuf,
-    shasum_path: &PathBuf,
-    binary_name: &str,
-) -> Result<(), ()> {
-    // see if file is readable and print out the first 64 bytes, which should be unique among them.
-    let file_read_sample = File::open(bin_path).expect("file to be readable");
-    let mut reader = BufReader::with_capacity(64, file_read_sample);
-    let bytes_read = reader.fill_buf().expect("reader to fill_buf");
-    println!("{:?} bytes : {:?}", &bin_path, bytes_read);
-
-    // fast, soft binary check
-    const LWD_BYTES: [u8; 64] = [
-        127, 69, 76, 70, 2, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 62, 0, 1, 0, 0, 0, 64, 188, 71,
-        0, 0, 0, 0, 0, 64, 0, 0, 0, 0, 0, 0, 0, 56, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 64, 0, 56, 0,
-        9, 0, 64, 0, 36, 0, 33, 0,
-    ];
-    const ZAI_BYTES: [u8; 64] = [
-        127, 69, 76, 70, 2, 1, 1, 3, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 62, 0, 1, 0, 0, 0, 0, 87, 19, 0,
-        0, 0, 0, 0, 64, 0, 0, 0, 0, 0, 0, 0, 112, 143, 238, 0, 0, 0, 0, 0, 0, 0, 0, 0, 64, 0, 56,
-        0, 14, 0, 64, 0, 34, 0, 33, 0,
-    ];
-    const ZCD_BYTES: [u8; 64] = [
-        127, 69, 76, 70, 2, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 62, 0, 1, 0, 0, 0, 0, 58, 121,
-        3, 0, 0, 0, 0, 64, 0, 0, 0, 0, 0, 0, 0, 8, 39, 154, 10, 0, 0, 0, 0, 0, 0, 0, 0, 64, 0, 56,
-        0, 12, 0, 64, 0, 47, 0, 45, 0,
-    ];
-    const ZCC_BYTES: [u8; 64] = [
-        127, 69, 76, 70, 2, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 62, 0, 1, 0, 0, 0, 208, 254, 85,
-        3, 0, 0, 0, 0, 64, 0, 0, 0, 0, 0, 0, 0, 216, 43, 87, 4, 0, 0, 0, 0, 0, 0, 0, 0, 64, 0, 56,
-        0, 12, 0, 64, 0, 47, 0, 45, 0,
-    ];
-    const ZINGO_BYTES: [u8; 64] = [
-        127, 69, 76, 70, 2, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 62, 0, 1, 0, 0, 0, 48, 151, 16,
-        0, 0, 0, 0, 0, 64, 0, 0, 0, 0, 0, 0, 0, 56, 16, 122, 4, 0, 0, 0, 0, 0, 0, 0, 0, 64, 0, 56,
-        0, 14, 0, 64, 0, 34, 0, 33, 0,
-    ];
-    const ZEBRA_BYTES: [u8; 64] = [
-        127, 69, 76, 70, 2, 1, 1, 3, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 62, 0, 1, 0, 0, 0, 208, 141, 33,
-        0, 0, 0, 0, 0, 64, 0, 0, 0, 0, 0, 0, 0, 152, 215, 66, 5, 0, 0, 0, 0, 0, 0, 0, 0, 64, 0, 56,
-        0, 14, 0, 64, 0, 34, 0, 33, 0,
-    ];
-
-    // const version strings for soft-confirming binaries when found
-    // lwd uses `version` instead of `--version`, therefore the call returns stderr
-    // zaino has updated to support `--version` but we are now serving an older commit
-    const VS_ZEBRAD: &str = "zebrad 2.1.0";
-    const VS_ZCASHD: &str = "Zcash Daemon version v6.1.0";
-    const VS_ZCASHCLI: &str = "Zcash RPC client version v6.1.0";
-    const VS_LWD: &str =
-        "Use \"lightwalletd [command] --help\" for more information about a command.";
-    const VS_ZAINOD: &str = "zainod [OPTIONS]";
-    const VS_ZINGOCLI: &str = "Zingo CLI 0.1.1";
-
-    let mut vs = Command::new(bin_path);
-    vs.arg("--version")
-        .stderr(Stdio::piped())
-        .stdout(Stdio::piped())
-        .output()
-        .expect("command with --version argument and stddout + stderr to be created");
-
-    let mut std_out = String::new();
-    let mut std_err = String::new();
-    vs.spawn()
-        .expect("vs spawn to work")
-        .stdout
-        .expect("stdout to happen")
-        .read_to_string(&mut std_out)
-        .expect("writing to buffer to complete");
-    vs.spawn()
-        .expect("vs spawn to work")
-        .stderr
-        .expect("stderr to happen")
-        .read_to_string(&mut std_err)
-        .expect("writing to buffer to complete");
-
-    match binary_name {
-        "lightwalletd" => {
-            if bytes_read == LWD_BYTES {
-                println!("lightwalletd bytes okay!");
-            } else {
-                fs::remove_file(bin_path).expect("bin to be deleted");
-                println!("binary {} removed!", binary_name);
-                return Err(());
-            }
-            if !std_err.contains(VS_LWD) {
-                panic!("expected LWD version string incorrect")
-            }
-            println!("lightwalletd version string okay!");
-        }
-        "zainod" => {
-            if bytes_read == ZAI_BYTES {
-                println!("zainod bytes okay!");
-            } else {
-                fs::remove_file(bin_path).expect("bin to be deleted");
-                println!("binary {} removed!", binary_name);
-                return Err(());
-            }
-            if !std_err.contains(VS_ZAINOD) {
-                panic!("expected Zainod version string incorrect")
-            }
-            println!("zainod version string okay!");
-        }
-        "zcashd" => {
-            if bytes_read == ZCD_BYTES {
-                println!("zcashd bytes okay!");
-            } else {
-                fs::remove_file(bin_path).expect("bin to be deleted");
-                println!("binary {} removed!", binary_name);
-                return Err(());
-            }
-            if !std_out.contains(VS_ZCASHD) {
-                panic!("ZCD version string incorrect")
-            }
-            println!("zcashd version string okay!");
-        }
-        "zcash-cli" => {
-            if bytes_read == ZCC_BYTES {
-                println!("Zcash-cli bytes okay!");
-            } else {
-                fs::remove_file(bin_path).expect("bin to be deleted");
-                println!("binary {} removed!", binary_name);
-                return Err(());
-            }
-            if !std_out.contains(VS_ZCASHCLI) {
-                panic!("ZCC version string incorrect")
-            }
-            println!("Zcash-cli version string okay!");
-        }
-        "zebrad" => {
-            if bytes_read == ZEBRA_BYTES {
-                println!("zebrad bytes okay!");
-            } else {
-                fs::remove_file(bin_path).expect("bin to be deleted");
-                println!("binary {} removed!", binary_name);
-                return Err(());
-            }
-            if !std_out.contains(VS_ZEBRAD) {
-                panic!("Zebrad version string incorrect")
-            }
-            println!("zebrad version string okay!");
-        }
-        "zingo-cli" => {
-            if bytes_read == ZINGO_BYTES {
-                println!("Zingo-cli bytes okay!");
-            } else {
-                fs::remove_file(bin_path).expect("bin to be deleted");
-                println!("binary {} removed!", binary_name);
-                return Err(());
-            }
-            if !std_out.contains(VS_ZINGOCLI) {
-                panic!("Zingo-cli version string incorrect")
-            }
-            println!("Zingo-cli version string okay!");
-        }
-        _ => println!("looked for unknown binary"),
-    }
-    println!("confirming {} hashsum against local record", binary_name);
-
-    // hashes for confirming expected binaries
-    let mut buf: BufReader<File> = BufReader::new(File::open(shasum_path).expect("shasum to open"));
-    let mut shasum_record = String::new();
-    buf.read_to_string(&mut shasum_record)
-        .expect("buffer to write into String");
-
-    if shasum_record.contains(binary_name) {
-        let hash = shasum_record
-            .split_whitespace()
-            .next()
-            .expect("shasum_record to be splittable");
-
-        // run sha512sum against file and return hex encoded String
-        let res = sha512sum_file(bin_path);
-        println!(
-            "found sha512sum of binary. asserting hash equality of local record {}",
-            shasum_record
-        );
-
-        if res != hash {
-            fs::remove_file(bin_path).expect("bin to be deleted");
-            return Err(());
-        }
-        println!(
-            "binary hash matches local record! Completing validation process for {}",
-            binary_name
-        );
-    }
-
-    Ok(())
-}
-
-async fn fetch_binary(bin_path: &PathBuf, binary_name: &str) {
-    // find locally committed cert for binary-dealer remote
-    let cert: Certificate = reqwest::Certificate::from_pem(
-        &fs::read(get_manifest_dir().join("cert/cert.pem")).expect("cert file to be readable"),
-    )
-    .expect("reqwest to ingest cert");
-    println!("cert ingested : {:?}", cert);
-
-    // Client deafult is idle sockets being kept-alive 90 seconds
-    let req_client = reqwest::ClientBuilder::new()
-        .connection_verbose(true)
-        .zstd(true)
-        .use_rustls_tls()
-        .tls_info(true)
-        .connect_timeout(Duration::from_secs(10)) // to connect // defaults to None
-        .read_timeout(Duration::from_secs(15)) // how long to we wait for a read operation // defaults to no timeout
-        .add_root_certificate(cert)
-        .build()
-        .expect("client builder to read system configuration and initialize TLS backend");
-
-    // reqwest some stuff
-    let asset_url = format!("https://zingolabs.nexus:9073/{}", binary_name);
-    println!("fetching from {:?}", asset_url);
-    let fetch_url = Url::parse(&asset_url).expect("fetch_url to parse");
-
-    let mut res = req_client
-        .get(fetch_url)
-        .send()
-        .await
-        .expect("Response to be ok");
-
-    // with create_new, no file is allowed to exist at the target location
-    // with .mode() we are able to set permissions as the file is created.
-    let mut target_binary: File = File::options()
-        .read(true)
-        .write(true)
-        .create_new(true)
-        .mode(0o100775)
-        .open(bin_path)
-        .expect("new binary file to be created");
-    println!(
-        "new empty file for {} made. write about to start!",
-        binary_name
-    );
-
-    // simple progress bar
-    let progress = ["/", "-", "\\", "-", "o"];
-    let mut counter: usize = 0;
-
-    while let Some(chunk) = res
-        .chunk()
-        .await
-        .expect("result to chunk ok.. *NOT A FAILED TRANSFER!")
-    {
-        target_binary
-            .write_all(&chunk)
-            .expect("chunk writes to binary");
-        print!(
-            "\rplease wait, fetching data chunks : {}",
-            progress[counter]
-        );
-        counter = (counter + 1) % 5;
-    }
-    println!("\nfile {} write complete!\n", binary_name);
-}
-
-fn sha512sum_file(file_path: &PathBuf) -> String {
-    let file_bytes = std::fs::read(file_path).expect("to be able to read binary");
-    let mut hasher = Sha512::new();
-    hasher.update(&file_bytes);
-    hex::encode(hasher.finalize())
-}
